--- SD-K9/src/app/home/home.page.html
@@ -1,93 +1,0 @@
-<style>
-  .ion-item {
-    text-align: center;
-  }
-</style>
-
-<ion-header>
-  <ion-toolbar>
-    <ion-title>
-        NavConU
-    </ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content>
-  <ion-segment (ionChange)="toggleCampus($event)" class="campusTgl">
-    <ion-segment-button value="sgw">
-      <ion-label>SGW</ion-label>
-    </ion-segment-button>
-    <ion-segment-button value="lyl">
-      <ion-label>Layola</ion-label>
-    </ion-segment-button>
-  </ion-segment>
-
-  <div #map id="map">
-  </div>
-
-  <ion-fab class="locate-me-btn-wrapper">
-    <ion-fab-button (ionFocus)="locateUser()" color="light" class="locate-me-btn"><img src="assets/images/map-pin.png" width="20px" /></ion-fab-button>
-  </ion-fab>
-</ion-content>
-
-<lib-ionic-pullup (onExpand)="footerExpanded()" (onCollapse)="footerCollapsed()" [(state)]="footerState" class="bottom-pullup-menu">
-  <ion-toolbar color="primary" class="toolbar" (click)="toggleFooter()" #ionDragFooter>
-      <ion-buttons slot="primary">
-        <ion-button (click)="clickedSearch()">
-        </ion-button>
-      </ion-buttons>
-      <ion-searchbar placeholder="Search for a place or classroom"></ion-searchbar>
-    </ion-toolbar>
-
-    <ion-content>
-      <ion-content>
-        <ion-list >
-          <ion-button (click)="openModal()" expand="block">
-            <ion-icon name="time"></ion-icon>
-              Shuttle Schedule
-          </ion-button>
-          <ion-button (click)="openModal1()"expand="block">
-            <ion-icon name="settings"></ion-icon>
-              App Settings
-          </ion-button>
-        </ion-list>
-<!-- FAVORITES -->
-  <ion-card>
-    <ion-card-header>
-      <ion-card-title>Favorites</ion-card-title>
-    </ion-card-header>
-    <ion-card-content>
-          <ion-grid>
-                <ion-row>
-                  <ion-col>
-                    <ion-button>SOEN-390</ion-button>
-                  </ion-col>
-                  <ion-col>
-                    <ion-button>SOEN-344</ion-button>
-                  </ion-col>
-                  <ion-col>
-                    <ion-button>COMP-335</ion-button>
-                  </ion-col>
-                </ion-row>
-          </ion-grid>
-        </ion-card-content>
-    </ion-card>
-
-<!-- RECENT EXAMPLE -->
-
-<ion-item>
-  <ion-label>Recent</ion-label>
-  <ion-input></ion-input>
-</ion-item>
-<ion-card>
-  <ion-card-header>
-
-    <ion-card-title>H building 534 <div class="ion-text-end">2 min</div></ion-card-title>
-    <ion-card-subtitle>Sir George Williams Campus</ion-card-subtitle>
-  </ion-card-header>
-
-</ion-card>
-
-      </ion-content>
-    </ion-content>
-</lib-ionic-pullup>
