--- SD-K9/src/app/home/home.page.scss
@@ -1,32 +1,0 @@
-#map{
-  position:absolute;
-  top:0;
-  bottom:0;
-  left:0;
-  right:0;
-}
-
-.campusTgl{
-  width:70%;
-  position:relative;
-  z-index:1;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 20px;
-}
-.locate-me-btn-wrapper{
-  bottom:100px;
-  right: 20px;
-  z-index:2;
-  
-  .locate-me-btn{
-    img{
-      pointer-events: none;
-    }
-  }
-}
-
-.bottom-pullup-menu{
-  z-index:99;
-  position:relative;
-}
